# Further optimise FastRegexMatcher